Move `register_function` into separate module to prevent circular imports.
